+++ clients/admin-ui/src/features/dataset/DataCategoryInput.tsx
@@ -12,14 +12,15 @@ import {
   Stack,
   Text,
 } from "@fidesui/react";
-import { useEffect, useMemo, useState } from "react";
+import { useMemo } from "react";
+
+import { DataCategory } from "~/types/api";
 
 import CheckboxTree from "../common/CheckboxTree";
 import { ArrowDownLineIcon } from "../common/Icon";
 import QuestionTooltip from "../common/QuestionTooltip";
 import DataCategoryTag from "../taxonomy/DataCategoryTag";
 import { transformDataCategoriesToNodes } from "../taxonomy/helpers";
-import { DataCategory } from "../taxonomy/types";
 
 interface Props {
   dataCategories: DataCategory[];
@@ -37,18 +38,6 @@ const DataCategoryDropdown = ({
     () => transformDataCategoriesToNodes(dataCategories),
     [dataCategories]
   );
-
-  const [internalChecked, setInternalChecked] = useState(checked);
-
-  useEffect(() => {
-    // make sure this stays up to date with the parent `checked` since this is not the only
-    // component that can control which items are checked
-    setInternalChecked(checked);
-  }, [checked]);
-
-  const handleClear = () => {
-    setInternalChecked([]);
-  };
 
   return (
     <Menu closeOnSelect>
@@ -82,11 +71,11 @@ const DataCategoryDropdown = ({
                   as={Button}
                   variant="outline"
                   size="xs"
-                  onClick={handleClear}
+                  onClick={() => onChecked([])}
                   mr={2}
                   width="auto"
                   closeOnSelect={false}
-                  data-testid="clear-btn"
+                  data-testid="data-category-clear-btn"
                 >
                   Clear
                 </MenuItem>
@@ -97,9 +86,8 @@ const DataCategoryDropdown = ({
                   colorScheme="primary"
                   color="white"
                   _hover={{ backgroundColor: "primary.600" }}
-                  onClick={() => onChecked(internalChecked)}
                   width="auto"
-                  data-testid="done-btn"
+                  data-testid="data-category-done-btn"
                 >
                   Done
                 </MenuItem>
@@ -110,8 +98,8 @@ const DataCategoryDropdown = ({
           <Box px={2}>
             <CheckboxTree
               nodes={dataCategoryNodes}
-              selected={internalChecked}
-              onSelected={setInternalChecked}
+              selected={checked}
+              onSelected={onChecked}
             />
           </Box>
         </Box>