+++ tests/core/test_api.py
@@ -10,9 +10,9 @@ from pytest import MonkeyPatch
 from starlette.testclient import TestClient
 
 from fidesctl.api.routes import health
+from fidesctl.api.routes.util import API_PREFIX
 from fidesctl.core import api as _api
 from fidesctl.core.config import FidesctlConfig
-from fidesctl.core.utils import API_PREFIX
 
 
 # Helper Functions